+++ flask_constance/backends/__init__.py
@@ -1,10 +1,6 @@
-from .core import Backend  # noqa
-from .memory import MemoryBackend  # noqa
+from .base import Backend, BackendCache  # noqa
+
 try:
     from .fsqla import FlaskSQLAlchemyBackend  # noqa
-except ImportError:
-    pass
-try:
-    from .redis import RedisBackend  # noqa
 except ImportError:
     pass