+++ flask_constance/backends/base.py
@@ -1,0 +1,24 @@
+import typing as t
+
+
+class Backend(t.Protocol):
+    """Base backend class."""
+
+    def get(self, name: str) -> t.Any:
+        ...
+
+    def set(self, name: str, value: t.Any):
+        ...
+
+
+class BackendCache(t.Protocol):
+    """Base backend cache class."""
+
+    def get(self, name: str) -> t.Any:
+        ...
+
+    def set(self, name: str, value: t.Any):
+        ...
+
+    def invalidate(self, name: str):
+        ...