+++ flask_constance/storage.py
@@ -1,11 +1,24 @@
 import typing as t
+from functools import partial
 
 from flask import current_app, g
+from contextlib import contextmanager
 
-from . import exc
-from .backends import Backend, MemoryBackend
-from .backends.exc import SettingNotFoundInBackendError
+from .backends import Backend, BackendCache
 from .signals import constance_get, constance_set
+
+
+@contextmanager  # type: ignore
+def _mut_context_manager(storage: "Storage", name: str) -> t.ContextManager[t.Any]:  # type: ignore
+    """Context manager for updating muttable values in storage.
+
+    :param storage: Storage.
+    :param name: Name of the setting.
+    :yields: Context manager returning value of setting.
+    """
+    value = getattr(storage, name)
+    yield value
+    setattr(storage, name, value)
 
 
 class Storage:
@@ -18,11 +31,17 @@ class Storage:
     """
 
     # NOTE: This hack prevents setting undeclared constance settings.
-    __slots__ = ["_backend", "_cache"]
+    __slots__ = ["_backend", "_cache", "_initialized"]
 
-    def __init__(self, backend: t.Optional[Backend] = None, cache: t.Optional[Backend] = None):
-        self._backend: Backend = backend or MemoryBackend()
-        self._cache: t.Optional[Backend] = cache
+    def __init__(self, backend: Backend, cache: t.Optional[BackendCache] = None):
+        self._backend: Backend = backend
+        self._cache: t.Optional[BackendCache] = cache
+        self._initialized = True
+
+    @property
+    def mut(self) -> t.Callable[[str], t.ContextManager[t.Any]]:
+        """Context manager to partial updating muttable settings."""
+        return partial(_mut_context_manager, self)
 
     @property
     def _payload(self) -> t.Dict[str, t.Any]:
@@ -37,25 +56,25 @@ class Storage:
         return g._constance_runtime_cache
 
     def __dir__(self) -> t.Iterable[str]:
-        return self._payload.keys()
+        return tuple(self._payload.keys())
 
     def __getattr__(self, name: str) -> t.Any:
         if name not in self._payload:
-            raise exc.ConstanceSettingNotFoundError(name)
+            return super().__getattribute__(name)
         constance_get.send(self, name=name)
         if name in self._ctx_cache:
             return self._ctx_cache[name]
         if self._cache is not None:
             try:
                 value = self._cache.get(name)
-            except SettingNotFoundInBackendError:
+            except KeyError:
                 pass
             else:
                 self._ctx_cache[name] = value
                 return value
         try:
             value = self._backend.get(name)
-        except SettingNotFoundInBackendError:
+        except KeyError:
             value = self._payload[name]
             self._backend.set(name, value)
             if self._cache is not None:
@@ -69,13 +88,12 @@ class Storage:
             return value
 
     def __setattr__(self, name: str, value: t.Any):
-        try:
-            if name not in self._payload:
-                return super().__setattr__(name, value)
-        except RuntimeError:  # NOTE: Prevent falling out while initializing extension.
+        if name in self.__slots__:
+            return super().__setattr__(name, value)
+        if name not in self._payload:
             return super().__setattr__(name, value)
         constance_set.send(self, name=name, value=value)
         self._backend.set(name, value)
         if self._cache is not None:
-            self._cache.set(name, value)
+            self._cache.invalidate(name)
         self._ctx_cache[name] = value