+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import random
+import typing as t
 from string import ascii_letters
 
 import pytest
@@ -6,44 +7,44 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
 from flask_constance import Constance
-from flask_constance.backends.core import Backend
+from flask_constance.storage import Storage
 from flask_constance.backends.fsqla import FlaskSQLAlchemyBackend, SettingMixin
 from flask_constance.backends.memory import MemoryBackend
-from flask_constance.backends.redis import RedisBackend
 
 
 def genstr(n: int = 8):
     return "".join(random.choice(ascii_letters) for _ in range(n))
 
 
+@pytest.fixture(scope="session")
+def default_payload():
+    return {
+        genstr(): genstr(),
+        genstr(): [genstr() for _ in range(random.randint(1, 10))],
+        genstr(): None,
+        genstr(): True,
+        genstr(): {genstr(): genstr()},
+    }
+
+
 @pytest.fixture
 def app():
     app = Flask(__name__)
     app.testing = True
-    Constance(app)
+    yield app
+
+
+@pytest.fixture
+def with_memory_backend(app: Flask, default_payload: t.Dict[str, t.Any]):
+    constance = Constance(app)
+    app.config["CONSTANCE_PAYLOAD"] = default_payload
     with app.app_context():
-        yield app
+        yield constance.storage._backend
 
 
 @pytest.fixture
-def dummy_backend():
-    class DummyBackend(Backend):
-        pass
-
-    backend = DummyBackend()
-    return backend
-
-
-@pytest.fixture
-def memory_backend():
-    backend = MemoryBackend()
-    return backend
-
-
-@pytest.fixture
-def fsqla_backend():
-    app = Flask(__name__)
-    app.testing = True
+def with_fsqla_backend(app: Flask, default_payload: t.Dict[str, t.Any]):
+    app.config["CONSTANCE_PAYLOAD"] = default_payload
     app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///:memory:"
     app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
 
@@ -53,18 +54,8 @@ def fsqla_backend():
         pass
 
     backend = FlaskSQLAlchemyBackend(Setting, db.session)
+    constance = Constance(app, backend)
 
     with app.app_context():
         db.create_all()
-        yield backend
-
-
-@pytest.fixture
-def redis_backend(redisdb):
-    app = Flask(__name__)
-    app.testing = True
-
-    backend = RedisBackend(redisdb)
-
-    with app.app_context():
         yield backend