+++ CHANGELOG.rst
@@ -13,7 +13,8 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0>`_
 
 First release with main functions of the package.
 
-Implemented backends:
+Features
+--------
 
-* Flask-SQLAlchemy
-* Redis
+- Implemented Flask-SQLAlchemy backend.
+- Implemented Memoty backend for testing.